+++ ext_emconf.php
@@ -8,11 +8,11 @@ $EM_CONF[$_EXTKEY] = [
     'author_company' => 'Causal Sàrl, Reelworx GmbH',
     'author_email' => 'xavier@causal.ch',
     'state' => 'stable',
-    'version' => '4.0.0',
+    'version' => '5.0.0',
     'constraints' => [
         'depends' => [
-            'php' => '8.1.99-8.4.99',
-            'typo3' => '11.5.0-13.4.99',
+            'php' => '8.4.0-8.5.99',
+            'typo3' => '13.4.0-14.3.99',
         ],
         'conflicts' => [],
         'suggests' => [],